requirements.txt
chromadb>=0.4.24
google-genai>=0.1.0
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
